Split into servant-multipart-api, servant-multipart-client, servant-multipart
