+++ servant-multipart-api/servant-multipart-api.cabal
@@ -1,4 +1,4 @@
-name:               servant-multipart
+name:               servant-multipart-api
 version:            0.12
 synopsis:           multipart/form-data (e.g file upload) support for servant
 description:
@@ -22,66 +22,20 @@ tested-with: GHC ==8.0.2 || ==8.2.2 || ==8.4.4 || ==8.6.5 || ==8.8.4 || ==8.10.2
 library
   default-language: Haskell2010
   hs-source-dirs:   src
-  exposed-modules:  Servant.Multipart
+  exposed-modules:  Servant.Multipart.API
 
   -- ghc boot libs
   build-depends:
-      array         >=0.5.1.1  && <0.6
-    , base          >=4.9      && <5
+      base          >=4.9      && <5
     , bytestring    >=0.10.8.1 && <0.11
-    , directory     >=1.3      && <1.4
     , text          >=1.2.3.0  && <1.3
     , transformers  >=0.5.2.0  && <0.6
-    , random        >=0.1.1    && <1.2
 
   -- other dependencies
   build-depends:
       http-media          >=0.7.1.3  && <0.9
-    , lens                >=4.17     && <4.20
     , resourcet           >=1.2.2    && <1.3
     , servant             >=0.16     && <0.19
-    , servant-client-core >=0.16     && <0.19
-    , servant-docs        >=0.10     && <0.19
-    , servant-foreign     >=0.15     && <0.19
-    , servant-server      >=0.16     && <0.19
-    , string-conversions  >=0.4.0.1  && <0.5
-    , wai                 >=3.2.1.2  && <3.3
-    , wai-extra           >=3.0.24.3 && <3.2
-
-executable upload
-  hs-source-dirs:   exe
-  main-is:          Upload.hs
-  default-language: Haskell2010
-  build-depends:
-      base
-    , bytestring
-    , http-client
-    , network            >=2.8 && <3.2
-    , servant
-    , servant-multipart
-    , servant-server
-    , servant-client
-    , servant-client-core
-    , text
-    , transformers
-    , wai
-    , warp
-
-test-suite servant-multipart-test
-  type:             exitcode-stdio-1.0
-  hs-source-dirs:   test
-  main-is:          Test.hs
-  default-language: Haskell2010
-  build-depends:
-      base
-    , bytestring
-    , http-types
-    , servant-multipart
-    , servant-server
-    , string-conversions
-    , tasty
-    , tasty-wai
-    , text
 
 source-repository head
   type:     git
