+++ servant-multipart-client/servant-multipart-client.cabal
@@ -1,0 +1,61 @@
+name:               servant-multipart-client
+version:            0.12
+synopsis:           multipart/form-data (e.g file upload) support for servant
+description:
+  This package adds support for file upload to the servant ecosystem. It draws
+  on ideas and code from several people who participated in the
+  (in)famous [ticket #133](https://github.com/haskell-servant/servant/issues/133) on
+  servant's issue tracker.
+
+homepage:           https://github.com/haskell-servant/servant-multipart#readme
+license:            BSD3
+license-file:       LICENSE
+author:             Alp Mestanogullari
+maintainer:         alpmestan@gmail.com
+copyright:          2016-2017 Alp Mestanogullari, 2018-2019 Servant Contributors
+category:           Web, Servant
+build-type:         Simple
+cabal-version:      >=1.10
+extra-source-files: CHANGELOG.md
+tested-with: GHC ==8.0.2 || ==8.2.2 || ==8.4.4 || ==8.6.5 || ==8.8.4 || ==8.10.2
+
+library
+  default-language: Haskell2010
+  hs-source-dirs:   src
+  exposed-modules:  Servant.Multipart.Client
+
+  -- ghc boot libs
+  build-depends:
+      array         >=0.5.1.1  && <0.6
+    , base          >=4.9      && <5
+    , bytestring    >=0.10.8.1 && <0.11
+    , directory     >=1.3      && <1.4
+    , text          >=1.2.3.0  && <1.3
+    , transformers  >=0.5.2.0  && <0.6
+    , random        >=0.1.1    && <1.2
+
+  -- other dependencies
+  build-depends:
+      servant-multipart-api == 0.12.*
+    , http-media          >=0.7.1.3  && <0.9
+    , servant             >=0.16     && <0.19
+    , servant-client-core >=0.16     && <0.19
+
+executable upload
+  hs-source-dirs:   exe
+  main-is:          Upload.hs
+  default-language: Haskell2010
+  build-depends:
+      base
+    , bytestring
+    , http-client
+    , network            >=2.8 && <3.2
+    , servant
+    , servant-multipart
+    , servant-multipart-client
+    , servant-client
+    , servant-client-core
+    , servant-server
+    , text
+    , transformers
+    , warp
