+++ servant-multipart/servant-multipart.cabal
@@ -1,0 +1,69 @@
+name:               servant-multipart
+version:            0.12
+synopsis:           multipart/form-data (e.g file upload) support for servant
+description:
+  This package adds support for file upload to the servant ecosystem. It draws
+  on ideas and code from several people who participated in the
+  (in)famous [ticket #133](https://github.com/haskell-servant/servant/issues/133) on
+  servant's issue tracker.
+
+homepage:           https://github.com/haskell-servant/servant-multipart#readme
+license:            BSD3
+license-file:       LICENSE
+author:             Alp Mestanogullari
+maintainer:         alpmestan@gmail.com
+copyright:          2016-2017 Alp Mestanogullari, 2018-2019 Servant Contributors
+category:           Web, Servant
+build-type:         Simple
+cabal-version:      >=1.10
+extra-source-files: CHANGELOG.md
+tested-with: GHC ==8.0.2 || ==8.2.2 || ==8.4.4 || ==8.6.5 || ==8.8.4 || ==8.10.2
+
+library
+  default-language: Haskell2010
+  hs-source-dirs:   src
+  exposed-modules:  Servant.Multipart
+
+  -- ghc boot libs
+  build-depends:
+      array         >=0.5.1.1  && <0.6
+    , base          >=4.9      && <5
+    , bytestring    >=0.10.8.1 && <0.11
+    , directory     >=1.3      && <1.4
+    , text          >=1.2.3.0  && <1.3
+    , transformers  >=0.5.2.0  && <0.6
+    , random        >=0.1.1    && <1.2
+
+  -- other dependencies
+  build-depends:
+      servant-multipart-api == 0.12.*
+    , http-media          >=0.7.1.3  && <0.9
+    , lens                >=4.17     && <4.20
+    , resourcet           >=1.2.2    && <1.3
+    , servant             >=0.16     && <0.19
+    , servant-docs        >=0.10     && <0.19
+    , servant-foreign     >=0.15     && <0.19
+    , servant-server      >=0.16     && <0.19
+    , string-conversions  >=0.4.0.1  && <0.5
+    , wai                 >=3.2.1.2  && <3.3
+    , wai-extra           >=3.0.24.3 && <3.2
+
+test-suite servant-multipart-test
+  type:             exitcode-stdio-1.0
+  hs-source-dirs:   test
+  main-is:          Test.hs
+  default-language: Haskell2010
+  build-depends:
+      base
+    , bytestring
+    , http-types
+    , servant-multipart
+    , servant-server
+    , string-conversions
+    , tasty
+    , tasty-wai
+    , text
+
+source-repository head
+  type:     git
+  location: https://github.com/haskell-servant/servant-multipart
